move components/legacy/utils/fs/remove-files-and-empty-dirs-recursively.ts to legacy
