+++ src/consumer/component/sources/remove-files-and-empty-dirs-recursively.ts
