+++ src/consumer/component/sources/remove-path.ts
@@ -1,5 +1,5 @@
 import fs from 'fs-extra';
-import { removeFilesAndEmptyDirsRecursively } from '@teambit/legacy.utils';
+import { removeFilesAndEmptyDirsRecursively } from './remove-files-and-empty-dirs-recursively';
 
 export default class RemovePath {
   path: string;
